Track mDNS initialization state on each netif num
LWIP assertion clearly shows that calling mdns_resp_add_netif twice for the same interface is a no-no.
So to prevent this we need to be able to easily check if this was already called.

This PR adds a simple map that keeps track of the initialization state per each interface (identified via netif->num).